fix(healthz): extract health module + fix sdNotify timing + await shutdown (#71)
- Move createHealthServer() to src/health.ts — pure side-effect-free module,
  prevents main() from running when tests import the factory
- sdNotify.ready() + startWatchdogMode() now fire inside grammy's onStart callback,
  so systemd READY= is sent only after long-polling has entered the update loop
- Await healthServer.close() in shutdown path to prevent DB/bot close racing open sockets
